\name{Quant_13B_problem_U}
\alias{Quant_13B_problem_U}
\docType{data}
\title{
   Test data for 24 math calculation questions from the SweSAT data.
}
\description{
   These data are for a randomly selected subset of 1000 examinees.
}
\usage{
  Quant_13B_problem_U
}
\format{
  A matrix object with 1000 rows and 24 columns.  The
  integers indicate which answer was chosen for each question by the examinee
  associated with the row.
}
\keyword{datasets}
