\name{Quantshort_key}
\alias{Quantshort_key}
\docType{data}
\title{
  Option information for the short form of the SweSAT Quantitative test.
}
\description{
  A vector that contains the indices of the right answers among the options
  for the 24 questions
}
\usage{
  Quantshort_key
}
\keyword{datasets}
