\name{SDS_U}
\alias{SDS_U}
\docType{data}
\title{
   Test data for Symptom Distress Scale.
}
\usage{
  SDS_U
}
\description{
The integers indicate which answer was chosen for each question by the examinee
  associated with the row.
}
\format{
  A vector object with 473 rows and 13 columns.  
}
\keyword{datasets}
