\name{Wbinsmth_nom}
\alias{Wbinsmth_nom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
List vector containing numbers of options and boundaries.
}
\description{
Set up objects needed for analyses of nominal data.  
}
\usage{
Wbinsmth_nom(bdry_nom, WfdList_nom)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bdry_nom}{Vector of length two containing the initial and final 
  values of the scofre index.
}
  \item{WfdList_nom}{A list vector of length equal to number of items.
  Each object is a list object containing the containing number of options
  and the nominal parameter matrix estimated by the mirt package.}
}
\details{
Called twice.  
}
