\name{Wbinsmth}
\alias{Wbinsmth}
\title{Estimate the option probability and surprisal curves.}
\description{The surprisal curves for each item are fit to the surprisal 
transforms of choice probabilities for each of a set of bins of current
performance values \code{theta}.  The error sums of squares are minimized
by the surprisal optimization \code{smooth.surp} in the \code{fda}
package.  The output is a list vector of length \code{n} containing the
functional data objects defining the curves.}
\usage{
  Wbinsmth(theta, dataList, WfdList=dataList$WfdList, 
                     thetaQnt=seq(0,100, len=2*nbin+1), wtvec=matrix(1,n,1),
                     iterlim=20, conv=1e-4, dbglev=0)
}
\arguments{
  \item{theta}{A vector of length N containing current values of score index 
    percentile values.}  
  \item{dataList}{A list that contains the objects needed to analyse the test or 
    rating scale.}
  \item{WfdList}{A numbered list object produced by a TestGardener analysis 
    of a test.  Its length is equal to the number of items in the test or 
    questions in the scale.  Each member of \code{WfdList} is a named list 
    containing information computed during the analysis.  By default 
    \code{Wbinsmth}  uses the \code{WfdList} member of \code{dataList}, 
    which is generated by the initialization with \code{Wbinsmth.init} within
    \code{make.dataList}.  Usually this is what is required.  
    However, \code{WfdList} may occasionally
    need to be entered from another source, and so included here as a separate 
    argument.
  }
  \item{thetaQnt}{A vector of length 2*n+1 containing the sequence of bin boundary and 
    bin centre values.}
  \item{wtvec}{A vector of length \code{n} of weights on observations.  Defaults to
    all ones.}
  \item{iterlim}{The maximum number of iterations used in optimizing surprisal
    curves.  Defaults to 20.}
  \item{conv}{Convergence tolerance.  Defaults to 0.0001.}
  \item{dbglev}{Level of output within \code{Wbinsmth}.  If 0, no output,
    if 1 the error sum of squares and slope on each iterations, and if 2
    or higher, results for each line search iteration with function \code{lnsrch}.}
}
\details{
  The function first bins the data in order to achieve rapid estimation of the 
  option surprisal curves. The argument \code{thetaQnt} contains the sequence 
  of bin boundaries separated by the bin centers, so that it is of length 
  \code{2*nbin + 1} where \code{nbin}
  is the number of bins.  
  These bin values are distributed over the percentile interval
  [0,100] so that the lowest boundary is 0 and highest 100.  
  Prior to the call to \code{Wbinsmth} these boundaries are computed so that 
  the numbers of values of \code{theta} falling in the bins are roughly equal.  
  It is important that the number of bins be chosen so that the bins contain 
  at least about 25 values.  
  
  After the values of \code{theta} are binned, the proportions that the bins 
  are chosen for each question and each option are computed.  Proportions of 
  zero are given NA values.
  
  The positive proportions are then converted to surprisal values where 
  surprisal = -log_M (proportion) where log_M is the logarithm with base M, 
  the number of options associated with a question.  
  Bins with zero proportions are assigned a surprisal that is appropriately 
  large in the sense of being in the range of the larger surprisal values 
  associated with small but positive proportions. This surprisal value
  is usually about 4.
  
  The next step is to fit the surprisal values for each question by a 
  functional data object that is smooth, passes as closely as possible to an
  option's surprisal values, and has values consistent with being a surprisal 
  value.  The function \code{smooth.surp()} is used for this purpose.  The
  arc length of thme item information curve is also computed.
  
  Finally the curves and other results for each question are saved in object
  \code{WfdList}, a list vector of length n, and the list vector is returned.
}
\value{
  The optimized numbered list object produced by \code{Wbinsmth}.  
  See above for its description.
}
\references{
Ramsay, J. O., Li J. and Wiberg, M. (2020) Full information optimal scoring. 
Journal of Educational and Behavioral Statistics, 45, 297-315.

Ramsay, J. O., Li J. and Wiberg, M. (2020) Better rating scale scores with 
information-based psychometrics. Psych, 2, 347-360.

http://testgardener.azurewebsites.net
}
\author{Juan Li and James Ramsay}
\seealso{
  \code{\link{ICC.plot}},
  \code{\link{Wbinsmth}}
}
\examples{
#  Example 1.  Display the item probability and surprisal curves for the 
#  short SweSAT multiple choice test with 24 items and 1000 examinees
#  estimate the initial surprisal curves
theta     <- Quant_13B_problem_dataList$percntrnk
thetaQnt  <- Quant_13B_problem_parList$thetaQnt
Qvec      <- Quant_13B_problem_parList$Qvec
WfdResult <- Wbinsmth(theta, Quant_13B_problem_dataList)
WfdList   <- WfdResult$WfdList
#  plot the curves for the first question
binctr    <- Quant_13B_problem_parList$binctr
scrfine   <- seq(0,100,len=101)
ICC.plot(scrfine, WfdList, Quant_13B_problem_dataList, Qvec, binctr, 
         plotindex=1, plotrange=c(0,100))
#  estimate the final surprisal curves
theta    <- Quant_13B_problem_parList$theta
thetaQnt <- Quant_13B_problem_parList$thetaQnt
Qvec     <- Quant_13B_problem_parList$Qvec
WfdList  <- Quant_13B_problem_parList$WfdList
#  plot the curves for the first question
ICC.plot(scrfine, WfdList, Quant_13B_problem_dataList, Qvec, binctr, 
         plotindex=1, plotrange=c(0,100))
#  Example 2.  Display the item probability and surprisal curves for the 
# Symptom Distress Scale with 13 items and 473 respondents.
# Proceed as above changing "Quant" to "SDS"
}
